PPU correctness fixes

+++ src/main.rs
@@ -72,7 +72,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
         game_state.update_joypad(a, b, start, select, up, down, left, right);
 
-        let cycles = cpu.step(&mut game_state);
+        let cycles = cpu.step(&mut game_state) as u16;
         let update = ppu.step(cycles, &mut game_state);
         // We unwrap here as we want this code to exit if it fails. Real applications may want to handle this in a different way
         if update {
